WARNING: I had to rename the inshell.h on the main directory for merging in ast directory. If needed, to compile: just rename the includes/inshell2.h to inshell.h and ast/inc/inshell.h to inshell2.h to avoid conflict.
Fixed make errors after merging the includes files.
Also deleted some functions from the files because already present in libft.

TODO: try to adapt exec system to work with an ast, or just give the result of the ast into a 2DArray in execution order to adapt parsing to exec.